- bug fixes to qsys master and slave (works with interconnect pipelining = 4 but not other numbers for some reason) - bug fixes to parse latency and queue time to have a fractional average latency instead of integer

+++ lynx/analysis/PerfAnalysis.java
@@ -318,7 +318,7 @@ public class PerfAnalysis {
             }
         }
 
-        int avgLatency = sumLatency / num;
+        double avgLatency = (double) sumLatency / num;
         return new LatencyStruct(connString(srcMod, dstMod), avgLatency, minLatency, maxLatency, num);
     }
 
@@ -357,7 +357,7 @@ public class PerfAnalysis {
             }
         }
 
-        int avgLatency = sumLatency / num;
+        double avgLatency = (double) sumLatency / num;
         return new LatencyStruct(connString(srcMod, dstMod), avgLatency, minLatency, maxLatency, num);
     }
